Disable history search + refactoring code
