+++ src/activeTab/uiEnhancer/messenger/updateNotes.js
